requirements.txt
Django>=5.2.4
requests
python-dotenv